+++ src/extension.ts
@@ -19,46 +19,128 @@ import { GenerateAssistantStubs } from './generators/generateAssistantStubs';
 import { InsertFunctionCall } from './generators/generateFunctionCall';
 import { DropProvider } from './providers/dropProvider';
 import { AdapterActionTreeView, Function } from './views/adapterActionTreeView';
-import { LoadConfigTreeView } from './views/loadConfigTreeView';
 import { AssistantByLabTreeView } from './views/assistantTreeView';
 import { ViewFileDecorationProvider } from './providers/decorationProvider';
-import { WorkflowTreeElement, WorkflowTreeView } from './views/workflowTreeView';
+import { WorkflowTreeView } from './views/workflowTreeView';
 import { ConfigTreeView } from './views/configTreeView';
 import { ConfigValues } from './providers/configProvider';
 import { initConfig } from './utils';
-import * as path from 'path';
-import { GenerateAdapterActionStubs } from './generators/generateAdapterActionStubs';
 import { ArtificialApollo } from './providers/apolloProvider';
 import { OutputLog } from './providers/outputLogProvider';
 import { WorkflowPublishLensProvider } from './providers/codeLensProvider';
+import { DataTreeView } from './views/dataTreeView';
 
 export async function activate(context: vscode.ExtensionContext) {
-  const rootPath =
-    vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0
-      ? vscode.workspace.workspaceFolders[0].uri.fsPath
-      : undefined;
-
+  // Config Setup
+  const { configVals, rootPath } = await setupConfig(context);
   if (!rootPath) {
     return;
   }
-  await initConfig(rootPath);
-  let devMode = false;
-  const configVals = ConfigValues.getInstance();
 
-  const outputLog = OutputLog.getInstance();
+  //Provides Type Error Decoration
+  new ViewFileDecorationProvider();
+  // Workflow Publishing Tree
+  new WorkflowTreeView(rootPath, context);
+  // Import/Export Buttons
+  new DataTreeView(rootPath, context);
+  // Config Tree
+  new ConfigTreeView(context);
+  // Commands to insert & drag/drop functions
+  setupDragAndDrop(context);
+  // Adapter Function Tree
+  const funcTree = setupAdapterFuncTree(configVals, context);
+  // Assistant Tree
+  const assistantByLab = await setupAssistantTree(configVals, context);
+  // Command to generate assistant stubs
+  new GenerateAssistantStubs(context, rootPath, assistantByLab);
+  //Drop handler for document editor
+  const selector = setupDropHandler(context, funcTree, assistantByLab);
+  // Status Bar for Connection Info
+  const statusBar = setupStatusBar(configVals, context);
+  // Code Lens for WF Publish
+  setupCodeLens(selector, context);
+  // Handle config resets across components
+  configResetWatcher(rootPath, configVals, statusBar, assistantByLab, context);
+  // Handle terminal command exit code notifications
+  taskExitWatcher();
+  console.log('Artificial Workflow Extension is active');
+}
 
-  const watchConfig = vscode.workspace.createFileSystemWatcher(
-    new vscode.RelativePattern(rootPath + '/configs', '**/*.yaml')
-  );
-
-  watchConfig.onDidChange((uri) => {
-    initConfig(rootPath);
+function taskExitWatcher() {
+  vscode.tasks.onDidEndTaskProcess((e) => {
+    if (e.exitCode === 0) {
+      vscode.window.showInformationMessage(`${e.execution.task.name} completed successfully`);
+    } else {
+      vscode.window.showErrorMessage(`${e.execution.task.name}: Please check terminal logs for error details`);
+    }
   });
-  context.subscriptions.push(watchConfig);
+}
 
+async function setupAssistantTree(configVals: ConfigValues, context: vscode.ExtensionContext) {
+  const assistantByLab = new AssistantByLabTreeView(
+    configVals.getAssistantStubPath(),
+    'artificial/assistantByLab/',
+    context
+  );
+  await assistantByLab.init();
+  return assistantByLab;
+}
+
+function setupAdapterFuncTree(configVals: ConfigValues, context: vscode.ExtensionContext) {
+  const funcTree = new AdapterActionTreeView(configVals.getAdapterActionStubPath(), context);
+  funcTree.init();
+  return funcTree;
+}
+
+function setupDragAndDrop(context: vscode.ExtensionContext) {
+  const funcCall = new InsertFunctionCall();
+  context.subscriptions.push(
+    vscode.commands.registerCommand('adapterActions.addToFile', (node: Function) => funcCall.insertFunction(node)),
+    vscode.commands.registerCommand('assistantsByLab.addToFile', (node: Function) => funcCall.insertFunction(node))
+  );
+}
+
+function setupDropHandler(
+  context: vscode.ExtensionContext,
+  funcTree: AdapterActionTreeView,
+  assistantByLab: AssistantByLabTreeView
+) {
+  const selector: vscode.DocumentFilter = { language: 'python', scheme: 'file' };
+  context.subscriptions.push(
+    vscode.languages.registerDocumentDropEditProvider(selector, new DropProvider(context, funcTree, assistantByLab))
+  );
+  return selector;
+}
+
+function setupCodeLens(selector: vscode.DocumentFilter, context: vscode.ExtensionContext) {
+  const codeLensProviderDisposable = vscode.languages.registerCodeLensProvider(
+    selector,
+    new WorkflowPublishLensProvider()
+  );
+  context.subscriptions.push(codeLensProviderDisposable);
+}
+
+function setupStatusBar(configVals: ConfigValues, context: vscode.ExtensionContext) {
+  const statusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
+  const host = configVals.getHost().split('.')[0];
+  statusBar.text = `$(debug-disconnect) ` + host;
+  statusBar.tooltip = `Artificial Workflow extension connected to ${configVals.getHost()}`;
+  statusBar.show();
+  context.subscriptions.push(statusBar);
+  return statusBar;
+}
+
+function configResetWatcher(
+  rootPath: string,
+  configVals: ConfigValues,
+  statusBar: vscode.StatusBarItem,
+  assistantByLab: AssistantByLabTreeView,
+  context: vscode.ExtensionContext
+) {
   const watchMergedConfig = vscode.workspace.createFileSystemWatcher(
     new vscode.RelativePattern(rootPath + '/tmp', 'merged.yaml')
   );
+  const outputLog = OutputLog.getInstance();
   watchMergedConfig.onDidChange((uri) => {
     outputLog.log('merged.yaml changed');
     configVals.reset();
@@ -67,116 +149,25 @@ export async function activate(context: vscode.ExtensionContext) {
     statusBar.text = `$(debug-disconnect) ` + configVals.getHost().split('.')[0];
     statusBar.tooltip = `Artificial Workflow extension connected to ${configVals.getHost()}`;
     assistantByLab.refresh();
-    loadConfigTree.refresh();
   });
   context.subscriptions.push(watchMergedConfig);
+}
 
-  vscode.commands.registerCommand('artificial-workflows.toggleDevMode', () =>
-    vscode.commands.executeCommand('setContext', 'devMode', (devMode = !devMode))
+async function setupConfig(context: vscode.ExtensionContext) {
+  const rootPath =
+    vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0
+      ? vscode.workspace.workspaceFolders[0].uri.fsPath
+      : '';
+  await initConfig(rootPath);
+  const configVals = ConfigValues.getInstance();
+  const watchConfig = vscode.workspace.createFileSystemWatcher(
+    new vscode.RelativePattern(rootPath + '/configs', '**/*.yaml')
   );
-
-  //Provides Type Error Decoration
-  new ViewFileDecorationProvider();
-
-  // Workflow Publishing view
-  const workflowTree = new WorkflowTreeView(rootPath, context);
-  vscode.commands.registerCommand('workflows.refreshEntry', () => workflowTree.refresh());
-  vscode.commands.registerCommand('workflows.publish', (path: string, workflowIDs: string[]) =>
-    workflowTree.publishWorkflow(path, workflowIDs)
-  );
-  vscode.commands.registerCommand('workflows.treePublish', (node: WorkflowTreeElement) =>
-    workflowTree.publishWorkflow(node.path, node.workflowIds)
-  );
-  vscode.commands.registerCommand('workflows.generateBinary', (node: WorkflowTreeElement) =>
-    workflowTree.generateWorkflow(node.path, false)
-  );
-  vscode.commands.registerCommand('workflows.generateJson', (node: WorkflowTreeElement) =>
-    workflowTree.generateWorkflow(node.path, true)
-  );
-
-  //Function Tree and related commands
-  const customAdapterActionStubPath =
-    vscode.workspace.getConfiguration('artificial.workflow.author').adapterActionStubPath;
-
-  const fullAdapterActionStubPath = path.join(rootPath, customAdapterActionStubPath);
-  const funcTree = new AdapterActionTreeView(fullAdapterActionStubPath, context);
-  funcTree.init();
-  vscode.commands.registerCommand('adapterActions.refreshEntry', () => funcTree.refresh());
-  const functionCallProvider = new InsertFunctionCall();
-  context.subscriptions.push(
-    vscode.commands.registerCommand('adapterActions.addToFile', (node: Function) =>
-      functionCallProvider.insertFunction(node)
-    )
-  );
-
-  //Load Config Tree and related commands
-  const loadConfigTree: LoadConfigTreeView = new LoadConfigTreeView(rootPath, context);
-  await loadConfigTree.init();
-  vscode.commands.registerCommand('loadConfigs.refreshEntry', () => loadConfigTree.refresh());
-
-  const customAssistantStubPath = vscode.workspace.getConfiguration('artificial.workflow.author').assistantStubPath;
-  const fullAssistantStubPath = path.join(rootPath, customAssistantStubPath);
-  //Assistant Tree and Commands
-  const assistantByLab: AssistantByLabTreeView = new AssistantByLabTreeView(
-    fullAssistantStubPath,
-    'artificial/assistantByLab/',
-    context
-  );
-  await assistantByLab.init();
-  await assistantByLab.refresh();
-  vscode.commands.registerCommand('assistantsByLab.refreshEntry', () => assistantByLab.refresh());
-  context.subscriptions.push(
-    vscode.commands.registerCommand('assistantsByLab.addToFile', (node: Function) =>
-      functionCallProvider.insertFunction(node)
-    )
-  );
-
-  //Config Tree and Commands
-  const configTree: ConfigTreeView = new ConfigTreeView(context);
-  vscode.commands.registerCommand('configs.refreshEntry', () => configTree.refresh());
-
-  // Generate Stubs
-  const generateAssistantsProvider = new GenerateAssistantStubs(rootPath, assistantByLab);
-  context.subscriptions.push(
-    vscode.commands.registerCommand('assistantsByLab.generateAssistantStubs', () =>
-      generateAssistantsProvider.generateAssistantStubsCommand()
-    )
-  );
-
-  // const generateAdapterActionsProvider = new GenerateAdapterActionStubs(rootPath, funcTree);
-  // context.subscriptions.push(
-  //   vscode.commands.registerCommand('adapterActions.generateAdapterStubs', () =>
-  //     generateAdapterActionsProvider.generateAdapterActionStubsCommand()
-  //   )
-  // );
-
-  //Drop handler for document editor
-  const selector: vscode.DocumentSelector = { language: 'python' };
-  context.subscriptions.push(
-    vscode.languages.registerDocumentDropEditProvider(selector, new DropProvider(funcTree, assistantByLab))
-  );
-
-  const statusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
-
-  const host = configVals.getHost().split('.')[0];
-  statusBar.text = `$(debug-disconnect) ` + host;
-  statusBar.tooltip = `Artificial Workflow extension connected to ${configVals.getHost()}`;
-  statusBar.show();
-  context.subscriptions.push(statusBar);
-
-  let docSelector = {
-    language: 'python',
-    scheme: 'file',
-  };
-
-  let codeLensProviderDisposable = vscode.languages.registerCodeLensProvider(
-    docSelector,
-    new WorkflowPublishLensProvider()
-  );
-
-  context.subscriptions.push(codeLensProviderDisposable);
-
-  console.log('Artificial Workflow Extension is active');
+  watchConfig.onDidChange((uri) => {
+    initConfig(rootPath);
+  });
+  context.subscriptions.push(watchConfig);
+  return { configVals, rootPath };
 }
 
 export function deactivate() {}